requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cssbeautifier==1.14.11
Django==4.2.10
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
djlint==1.34.1
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.11
json5==0.9.14
pathspec==0.12.1
pillow==10.2.0
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.4
